Alter text size of the display based on content
Changing the font size of the display depending on
the content prevents the text from breaking to a
new line. So, when the user gets an error by
dividing by 0, the font gets smaller and is
centered to the left, and when the user enters
numbers, the font gets bigger and is centered to
the right.